+++ man-roxygen/engagements.R
@@ -1,1 +1,1 @@
-#' @param engagements An engagements object, typically produced by running `hs_engagements_raw()`
+#' @param engagements An engagements return from `hs_engagements_raw()`