+++ man-roxygen/lineitems.R
@@ -1,0 +1,1 @@
+#' @param lineitems A line items return from `hs_lineitems_raw()`