+++ man-roxygen/object_properties.R
@@ -1,0 +1,1 @@
+#' @param object_properties Object properties as returned by `hs_object_properties_raw()`