+++ man-roxygen/objectname.R
@@ -1,0 +1,1 @@
+#' @param objectname String name of the object. Can be one of `"tickets"`, `"products"` or `"line-items"`.