+++ man-roxygen/offsetvalue.R
@@ -1,0 +1,1 @@
+#' @param offsetvalue The offset value to use if continuing a previous call. Defaults to 0.