+++ man-roxygen/products.R
@@ -1,0 +1,1 @@
+#' @param products A products return from `hs_products_raw()`