+++ man-roxygen/tickets.R
@@ -1,0 +1,1 @@
+#' @param tickets A tickets return from `hs_tickets_raw()`